Extract MakeRightDomainOperationImpl from DomainProcessorImpl
