+++ operator/src/main/java/oracle/kubernetes/operator/helpers/JobHelper.java
@@ -178,14 +178,20 @@ public class JobHelper {
    *  ProcessingConstants.DOMAIN_RESTART_VERSION - the restart version from the domain
    *  ProcessingConstants.DOMAIN_INPUTS_HASH
    *  ProcessingConstants.DOMAIN_INTROSPECT_VERSION - the introspect version from the old domain spec
-   *
-   * @param next Next processing step
    */
+  public static Step createIntrospectionStartStep() {
+    return new IntrospectionStartStep();
+  }
+
   public static Step createIntrospectionStartStep(Step next) {
     return new IntrospectionStartStep(next);
   }
 
   private static class IntrospectionStartStep extends Step {
+
+    IntrospectionStartStep() {
+      super();
+    }
 
     IntrospectionStartStep(Step next) {
       super(next);
@@ -439,7 +445,7 @@ public class JobHelper {
 
     private Step createIntrospectionSteps(Step next) {
       return Step.chain(
-              readExistingIntrospectorConfigMap(getNamespace(), getDomainUid()),
+              readExistingIntrospectorConfigMap(),
               startNewIntrospection(next));
     }
