--- operator/src/main/java/oracle/kubernetes/operator/steps/DomainPresenceStep.java
@@ -1,26 +1,0 @@
-// Copyright (c) 2017, 2022, Oracle and/or its affiliates.
-// Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
-
-package oracle.kubernetes.operator.steps;
-
-import oracle.kubernetes.operator.work.NextAction;
-import oracle.kubernetes.operator.work.Packet;
-import oracle.kubernetes.operator.work.Step;
-import oracle.kubernetes.weblogic.domain.model.DomainResource;
-
-public class DomainPresenceStep extends Step {
-
-  private DomainPresenceStep(Step domainUpSteps) {
-    super(domainUpSteps);
-  }
-
-  public static DomainPresenceStep createDomainPresenceStep(
-      DomainResource dom, Step domainUpSteps, Step managedServerStep) {
-    return new DomainPresenceStep(dom.isShuttingDown() ? managedServerStep : domainUpSteps);
-  }
-
-  @Override
-  public NextAction apply(Packet packet) {
-    return doNext(packet);
-  }
-}
