+++ operator/src/test/java/oracle/kubernetes/operator/DomainProcessorTest.java
@@ -276,7 +276,7 @@ class DomainProcessorTest {
 
   @Test
   void whenDomainSpecNotChanged_dontRunUpdateThread() {
-    DomainProcessorImpl.registerDomainPresenceInfo(new DomainPresenceInfo(newDomain));
+    processor.registerDomainPresenceInfo(new DomainPresenceInfo(newDomain));
 
     makeRightOperation.execute();
 
@@ -291,7 +291,7 @@ class DomainProcessorTest {
   
   @Test
   void whenDomainExplicitSet_runUpdateThread() {
-    DomainProcessorImpl.registerDomainPresenceInfo(new DomainPresenceInfo(domain));
+    processor.registerDomainPresenceInfo(new DomainPresenceInfo(domain));
 
     processor.createMakeRightOperation(new DomainPresenceInfo(domain)).withExplicitRecheck().execute();
 
@@ -300,7 +300,7 @@ class DomainProcessorTest {
 
   @Test
   void whenDomainChangedSpec_runUpdateThread() {
-    DomainProcessorImpl.registerDomainPresenceInfo(new DomainPresenceInfo(domain));
+    processor.registerDomainPresenceInfo(new DomainPresenceInfo(domain));
 
     processor.createMakeRightOperation(new DomainPresenceInfo(newDomain)).execute();
 
@@ -309,7 +309,7 @@ class DomainProcessorTest {
 
   @Test
   void whenDomainChangedSpecNewer_setWillInterrupt() {
-    DomainProcessorImpl.registerDomainPresenceInfo(new DomainPresenceInfo(domain));
+    processor.registerDomainPresenceInfo(new DomainPresenceInfo(domain));
 
     final MakeRightDomainOperation operation = processor.createMakeRightOperation(new DomainPresenceInfo(newDomain));
 
@@ -318,7 +318,7 @@ class DomainProcessorTest {
 
   @Test
   void whenDomainChangedSpecNotNewer_dontSetWillInterrupt() {
-    DomainProcessorImpl.registerDomainPresenceInfo(new DomainPresenceInfo(newDomain));
+    processor.registerDomainPresenceInfo(new DomainPresenceInfo(newDomain));
 
     final MakeRightDomainOperation operation = processor.createMakeRightOperation(new DomainPresenceInfo(domain));
 
@@ -327,7 +327,7 @@ class DomainProcessorTest {
 
   @Test
   void whenDomainChangedSpecButProcessingAborted_dontRunUpdateThread() {
-    DomainProcessorImpl.registerDomainPresenceInfo(new DomainPresenceInfo(domain));
+    processor.registerDomainPresenceInfo(new DomainPresenceInfo(domain));
     newDomain.getOrCreateStatus().addCondition(new DomainCondition(FAILED).withReason(ABORTED).withMessage("ugh"));
 
     processor.createMakeRightOperation(new DomainPresenceInfo(newDomain)).execute();
@@ -337,7 +337,7 @@ class DomainProcessorTest {
 
   @Test
   void whenDomainChangedSpecAndProcessingAbortedButRestartVersionChanged_runUpdateThread() {
-    DomainProcessorImpl.registerDomainPresenceInfo(new DomainPresenceInfo(domain));
+    processor.registerDomainPresenceInfo(new DomainPresenceInfo(domain));
     newDomain.getOrCreateStatus().addCondition(new DomainCondition(FAILED).withReason(ABORTED).withMessage("ugh"));
     domainConfigurator.withRestartVersion("17");
 
@@ -348,7 +348,7 @@ class DomainProcessorTest {
 
   @Test
   void whenDomainChangedSpecAndProcessingAbortedButInspectionVersionChanged_runUpdateThread() {
-    DomainProcessorImpl.registerDomainPresenceInfo(new DomainPresenceInfo(domain));
+    processor.registerDomainPresenceInfo(new DomainPresenceInfo(domain));
     newDomain.getOrCreateStatus().addCondition(new DomainCondition(FAILED).withReason(ABORTED).withMessage("ugh"));
     domainConfigurator.withIntrospectVersion("17");
 
@@ -359,7 +359,7 @@ class DomainProcessorTest {
 
   @Test
   void whenDomainChangedSpecAndProcessingAbortedButImageChanged_runUpdateThread() {
-    DomainProcessorImpl.registerDomainPresenceInfo(new DomainPresenceInfo(domain));
+    processor.registerDomainPresenceInfo(new DomainPresenceInfo(domain));
     newDomain.getOrCreateStatus().addCondition(new DomainCondition(FAILED).withReason(ABORTED).withMessage("ugh"));
     domainConfigurator.withDefaultImage("abcd:123");
 
@@ -1187,7 +1187,7 @@ class DomainProcessorTest {
     for (Integer i : msNumbers) {
       defineServerResources(getManagedServerName(i));
     }
-    DomainProcessorImpl.registerDomainPresenceInfo(new DomainPresenceInfo(domain));
+    processor.registerDomainPresenceInfo(new DomainPresenceInfo(domain));
     testSupport.defineResources(createIntrospectorConfigMap(OLD_INTROSPECTION_STATE, clusterNames, independentServers));
     testSupport.doOnCreate(KubernetesTestSupport.JOB, j -> recordJob((V1Job) j));
     domainConfigurator.withIntrospectVersion(OLD_INTROSPECTION_STATE);
@@ -1437,7 +1437,7 @@ class DomainProcessorTest {
   private void getMIIOnlineUpdateIntrospectResult(DomainConditionType domainConditionType, String updateResult)
       throws Exception {
 
-    DomainProcessorImpl.registerDomainPresenceInfo(new DomainPresenceInfo(domain));
+    processor.registerDomainPresenceInfo(new DomainPresenceInfo(domain));
 
     String introspectorResult = ">>>  /u01/introspect/domain1/userConfigNodeManager.secure\n"
         + "#WebLogic User Configuration File; 2\n"
@@ -1978,7 +1978,7 @@ class DomainProcessorTest {
   @Test
   void whenExceptionDuringProcessing_reportInDomainStatus() {
     DomainPresenceInfo info = new DomainPresenceInfo(domain);
-    DomainProcessorImpl.registerDomainPresenceInfo(info);
+    processor.registerDomainPresenceInfo(info);
     forceExceptionDuringProcessing();
 
     testSupport.setTime(DomainPresence.getDomainPresenceFailureRetrySeconds(), TimeUnit.SECONDS);
@@ -1996,7 +1996,7 @@ class DomainProcessorTest {
   @Test
   void whenExceptionDuringProcessing_createFailedEvent() {
     DomainPresenceInfo info = new DomainPresenceInfo(domain);
-    DomainProcessorImpl.registerDomainPresenceInfo(info);
+    processor.registerDomainPresenceInfo(info);
     forceExceptionDuringProcessing();
 
     testSupport.setTime(DomainPresence.getDomainPresenceFailureRetrySeconds(), TimeUnit.SECONDS);
@@ -2013,7 +2013,7 @@ class DomainProcessorTest {
   @Test
   void whenWebLogicCredentialsSecretRemoved_NullPointerExceptionAndAbortedEventNotGenerated() {
     consoleHandlerMemento.ignoreMessage(NOT_STARTING_DOMAINUID_THREAD);
-    DomainProcessorImpl.registerDomainPresenceInfo(new DomainPresenceInfo(domain));
+    processor.registerDomainPresenceInfo(new DomainPresenceInfo(domain));
     domain.getSpec().withWebLogicCredentialsSecret(null);
     int time = 0;
